Bal Masque release
From EC:
Sessions 20, 24, 26, 31 March, 1 April, 1958.

The "Duke Ellington at the Bal Masque" has been published in CD on
Columbia COL 469136 2 and Collectables COL-CD-7856 (Duke’s Mixture / At
The Bal Masqué). (And, I think, on the boxset Complete Columbia Studio
Recordings 1951-58.)

The content of COL 469136 2 and COL-CD-7056 is the same of the stereo LP
Co CS 8098, and not of the mono Co CL 1282: Indian Love Call 58032603
and Alice Blue Gown 58033103.