Fix Lullaby Of Birdland releases (from EC)
58042421  CO60902-6  Lullaby Of Birdland  FMRS FM-4001/02                DE5819u
58042422  CO60902-7  Lullaby Of Birdland  FMRS FM-4001/02, CBS 88653(D)  DE5819v

58042421 should have a '*' (incomplete) for FMRS FM-4001/02, and *CBS 88653 should be added too.
Hence it should be
58042421  CO60902-6  Lullaby Of Birdland  FMRS FM-4001/02*, CBS 88653*(D)  DE5819u

Sources:
NDCS 5001: DE5819u. Add: (CBS 88653).
Brackets denote an excerpt, inserted on the main take

DEMS 04/3-31
Columbia/Legacy COL 512919 2. Duke Ellington — Piano in the Background
Duke had been struggling with this piano introduction earlier in the
Columbia recording session of 24Apr58, when he started the tune 7 times.
The second attempt has been released on the LP Up to Date 2007 and a
combination of the 6th and 7th take has been used for the LP Franklin
Mint 4002 and for the French (blue) CBD LP 88653 but none of it has been
re-released on CD.

DEMS 85/1-8 (re: CBS 88653): "Lullaby is the same as on Franklin Mint, thus different to the version on Up-To-Date 2007".