Removed asterix
Comments from EC:

You have raised a very good point. I have listened to the DVD and the CD
to compare Low Key Light.

In the CD Low Key Lightly is composed of two parts, the piano intro and
the violin + orchestral music. I assume they are respectively is
DE5920xb and DE5919b. In the DVD, only the piano intro is included
DE5920xb. Both in the CD and in the DVD their duration is similar,
approx. 1'20". I can't say if they are exactly the same piece of music,
because the dialogues in the DVD overwhelm the music.

I think that possibly brackets in NDESOR mean an insert rather than a
fragment: (5920xb) is inserted before DE5919b. In the description of The
CD CK 65569 (NDCS 3001), DE5920xb is not mentioned.

I wouldn't put the asterisk on CoT E-10071 here, because I cannot say if
the soundtrack contains the whole take DE5920xb or just a part. After
all, most of the pieces of music in a soundtrack are not complete.