Add Jane Harvey session
From EC:

A very rare Ellington recording

DEMS 12/2-15

Alan Eichler was so kind as to send us a set of five CDs with recordings
made by Jane Harvey. They were released by Little Jazz Bird Records.
They include newly discovered performances with Duke Ellington and his
orchestra with Billy Strayhorn, recorded in the late 1950's.

The two tracks with the Ellington band are on the fifth CD, titled "The
Undiscovered Jane Harvey", LJB 1005 (2011). The titles are A Hundred
Dreams from Now (originally titled A Hundred Dreams Ago) and I Ain't Got
Nothing But the Blues. It is unmistakably the Ellington orchestra with
Billy Strayhorn at the piano and Duke supervising the session and giving
instructions to Jane at the start of the second selection. The same
recording of A Hundred Dreams from Now is also on track 2 of the first
CD in the series, LJB 1001, titled "Travelin' Light".

We have asked Alan Eichler to give us the date and the location of these
recordings in order to include them in Duke's discography. He answered:
"The Ellington recordings were done in New York during Jane's marriage
to Bob Thiele. It would have become a full album, but Jane and Bob
divorced and she thought all copies had been destroyed. She estimates
the year as around 1959. DEMS