Changed Fiddle on the Diddler to Fiddler on the Diddle.
From DP:

This was a Reprise recording session at Fine Studios in New York according to New Desor, MacHare and Nielsen).

The title of Fiddle on the Diddler may need tweaking or an explanatory note.  DEMS http://tdwaw.ellingtonweb.ca/DEMS/DEMS-04-1d,200404-200407,pts30-34.html

'The 5 CD set of Mosaic and all my LPs have Fiddler on the Diddle.  Timner and Nielsen have Fiddle on the Diddle.

You are surely correct in querying this spelling.  It must be Fiddler on the Diddle.  The Leonard Feather sleeve notes to the original LP issue indicate that this is so, where Ellington himself is quoted, calling it an instance of 'using word-plays on the titles of various Broadway shows'.  The Fiddler on the Diddle spelling is the only one which retains the punning reference to Fiddler on the Roof.  Fiddle on the Diddle makes a kind of sense but loses the punning reference; Fiddle on the Diddler as per DE6533 doesn't make sense.  The Stanley Dance notes on the same LP issue though, have Fiddle on a Diddle, in which the punning reference is also lost.’

(It’s hard to tell if this is all Roger Boyes’ note or if part of it is Sjef Hoefsmit’s)

The New Desor authors agreed to the change.

401 - Session 6533, 6533x & y: change the title into Fiddler on the Diddle. (04/1-31)

(see http://tdwaw.ellingtonweb.ca/DEMS/DEMS-05-1e,200504-200507,pts38-41.html )

It’s your call.  I don’t need a reply.