Corrected "April 1965" Mike Douglas session to 26 March 1965 and clarified description.
From DP:

This one’s kind of a big correction.  New Desor had it wrong, and Marcus relied on New Desor.

This session was March 26, 1965 in Cleveland rather than Chicago April 19, and it wasn’t a CBS telecast although it probably played on a CBS station.

From 1961 to the middle of 1965, the Mike Douglas Show was a Westinghouse
Broadcasting Company Group W production, initially on the one Cleveland
station, then after negotiations with AFTRA, four more Group W stations across
the country.  By April 1965, it was syndicated and running on 43 stations.
This was Group W’s first syndication.
https://en.wikipedia.org/wiki/The_Mike_Douglas_Show  Some years earlier
Westinghouse was allegedly coerced by NBC to move its Philadelphia station to
Cleveland so NBC could have an outlet in Philly.   Legal action resulted in
this move being reversed in mid-1965 so the station relocated back to
Philadelphia in either June or August 1965, depending on which source you read.