Corrected Menelik - The Lion Of Judah (which has only one take). #88
erack 10 Menelik - The Lion Of Judah (4117c in NDESOR) and 11-Menelik - The
Lion Of Judah (4117xa in NDESOR) come from the same take. Therefore the NDESOR
Id 4117xa should be removed.

There was only one take of Menelik The Lion Of Judah: the so called Take 2,
identified as PBS 061344-2, is indeed a test pressing of the same Take 1,
PBS-061344-1. Examination of two master-pressed tests, marked "-1" and "-2"
shows that they derive from different master-cuttings, with very slightly
different run-off configurations (neither identical to HMV JO.282), a different
run-on configuration, and different volume levels. Playing the two "takes" back
simultaneously revealed that both are nonetheless the same take... Steven
Lasker. (DEMS 03/2-20/5).