Split up DE7344d into two parts (-1 and -2), in line with the presentation in the Centennial Edition
From EC:

As NDESOR has assigned one ID to the two parts of "Is God A Three-Letter Word For Love?" of the THIRS SACRED CONCERT, in ellingtonia there is one line only:

73-10-24-004	07. Is God A Three-Letter Word For Love? - vABa,JAC 	DE7344d

However on the LP and in the disc 20 of the RCA Centennial boxset, the piece is split in two parts, having distinct matriculates number (not present in ellingtonia): DPE6-1200 for part 1 and DPE6-3145 for part 2.
In the RCA 09026-63386-2 release page, you list only part 2, as track 7 of CD 20, with its duration (3:45).

73-10-24-004	(20-7) 07. Is God A Three-Letter Word For Love? - vABa,JAC (3:45) DE7344d

I think you should amend like this

73-10-24-004	(20-6,7) 07. Is God A Three-Letter Word For Love? - vABa,JAC (8:12) DE7344d

where 8:12 is the total time pf part 1 (4:27) and part 2 (3:45).

Otherwise, you might choose to split the song in two lines in the discography.