Restored ordering (but not NDESORS) for DE2812a-c
MG:

If I remember correctly, I discovered that the correct playing order was
different from the numbering in DESOR so I listed them in the order of playing
rather than the DESOR order.

EC:

I understand from Marcus that the original order in ellingtonia was right
(indeed the same order is on David Palmquist's website), and the NDESOR IDs are
not in the right order. The DESOR numbers are taken from the NDCS 3002, an
official document written or approved by the NDESOR authors.

So the resulting sequence should be:
BVE47799-2  The Mooche                                  DE2812c
BVE48100-1  Santa Claus,  Bring My Man Back To Me       DE2812a
BVE48101-2  I Done Caught You Blues                     DE3812b
Is it your understanding too?

On the other side, in the booklet of the RCA Centennial edition (I found the
pdf in internet, I don't have the boxset) there is the following,

1:45 to 3:00 Tuesday, October 30, 1928
18. Santa Claus,  Bring My Man Back To Me
19. I Done Caught You Blues

3:00 to 5:00' Tuesday, October 30, 1928
20. The Mooche

And I believe this is the source of the DESOR numbering,