Corrected drummer and location for 23 Jan 1966 session
From EC:

I've found that some correction is needed on Jan 23, 1966: the location is New
York and not Worcester, MA. The telecast is "The Ed Sullivan Show". The drummer
is not Louie Bellson: "Session 6603, 23Jan66. According to Ken Vail p279, the
drummer was Skeets Marsh. This makes sense. Klaus Stratemann stated on p528
that Louie Bellson was called back for the soundtrack recordings for the
picture "Assault on a Queen". Ken Vail wrote: "Immediately after the show
[23Jan66], Ellington and the band fly to Lisbon…." We know that Louie did not
go with the band to Europe and was replaced by Skeets Marsh,(DEMS
Bulletin 05/1-38). 

Actually, watching the video, the drummer is Afro-American, definitely not
Bellson.