Renamed El '97 to the commerical release JUCD-2069, and added one take to MDD 012
From EC

10 June
The CD El'97 (a souvenir of The Ellington conference in Leeds in 1997), has been
issued by Jazz Unlimited in 1999 (JUCD-2069). This is a commercial release,
therefore I would replace El'79 with JUCD-2069, or at least add this last one.

For completeness,
40-06-10-009	The Sergeant Was Shy
is also in Maison Du Duke CD 12 (MDD 012).