Added missing tracks to RCA 07863-66790-2 (and corrected a previous typo)
From EC

In ellingtonia, there are only the following 5 tracks in the release RCA 07863-66790-2 (Ben Webster . Cotton Tail).
11 Mobile Bay     BS 053609-1 2Nov40 DE4033j
12 Linger Awhile  BS 053610-2 2Nov40 DE4033m
15 Some Saturday  PBS-061342-1 3Jul41 DE4117a
19 Perdido           BS-070682-2 21Jan42 DE4201b
21 Main Stem      PBS-072438-1 26Jun42 DE4208b

This release was relevant, because it contained Perdido take 2 for the first
time on CD. Assuming that when mentioning a release, it should be best to
include all its Ellington's titles, the following are missing in the ellingtonia
release.

1  Cotton Tail             PBS 049655-1 4May40 DE4008b
2  All Too Soon.          BS 054608-1 22Jul40 DE4012c
7  Conga Brava           BS 049015-1 15Mar40 DE4006a
8  Bojangles-1            BS 053021-1 28May40 DE4009c
9  My Greatest Mistake BS 054624-1 24Jul40 DE4013a
10 Chloe                   BS 053580-1 28oct40 DE4032c
13 Blue Serge            PBS-055286-1 15Feb41 DE4105e
14 Just A-Sittin' And A-Rockin'  PBS-061285-1 5Jun41 DE4110d
16 Five O'Clock Drag        PBS-061684-1 26Sep41 DE4121a
17 Chelsea Bridge            PBS-061687-2 2Dec41 DE4125e
18 Rain Check                 PBS-061941-1 2Dec41 DE4125a
20 What Am I Here For?   BS-071890-1 26Feb42 DE4202a

Indeed in Ellingtonia.com the track 16 
41-09-26-001 BS061684-1 Five O'Clock Drag DE4121a
is assigned to the release RCA 07863-33790-2, that sounds like a mistake for RCA 07863-66790-2. You may want to correct.

For your info, the remaining tracks of the CD are not by Ellington, and are to be ignored. Here the details:
3  Toby (Bennie Moten's Kansas City Orchestra) 74847-1 13Dec32
4  Lafayette (Bennie Moten's Kansas City Orchestra) 74853-1 13Dec32
5  The Voice Of Old Man River (Willie Bryant And His Orchestra) 92908-1 1Aug35
6  Early Session Hop (Lionel Hampton And His Orchestra) 041409-1 11Sep39
22 Cadillac Slim (Benny Carter And His Chocolate Dandies) 2697 23Aug46