Deleted non-existent session DE6417 (the material was recorded in Goteborg on March 11, 1964)
From EC:

I hope you are doing well.
I am trying to help Charlie giving him input to updat and amend the
Ellingtonia.com website.

I have encountered in ellingtonia the 2nd Concert of March 9, 1964 in Stockholm.
In May 2000 you published a couple of excellent articles on the DESS web, where
you stated that the music that had been thought to come from that concert was
indeed recorded in Goteborg on March 11, 1964.

Does it mean that the concert DE6417 of NDESOR should be deleted? This is my
understanding, but I would appreciate your valuable confirmation.

Ulf:
Dear Ezio, sorry for this late replay. Yes, DE6417 of NDESOR should be deleted?
There is a list of all Ellington concerts in Sweden on the DESS website.