Changed vocalist for DE6363e/f to Jimmy Grissom
From EC:

Based on Brian Koller's advice, the following corrections should be done in ellingtonia.

21 August 1963.
Add Jimmy Grissom to the vocalists of the band.

And replace JMP with JG in the first two takes of "My Mother, My Father And Love" as below.
63-08-21-005 (05) 06. My Mother, My Father And Love - vJG	DE6363e
63-08-21-006 (06) 06. My Mother, My Father And Love - vJG	DE6363f
63-08-21-007 (07) 06. My Mother, My Father And Love - vJMP DE6363g
63-08-21-008 (08) 06. My Mother, My Father And Love - vJMP DE6363h

Brian Koller wrote :
«By the way, there is an error in the New Desor and ellingtonia.com for entry
DE6363f from August 21, 1963.  It is Jimmy Grissom instead of Jimmy McPhail on
"My Mother, My Father And Love."» (July 5, 2023)

«According to the New Desor, 6363e and 6363g are incomplete takes. I don't have
those recordings, which are available only on the unedited studio tape (from
August 21, 1963) in the Ellington stockpile held (?) by Radio Denmark. Copies of
that tape may be circulating. I am sure Sjef Hoefsmit had a copy, as did the New
Desor authors.  I should add that 6363e is a breakdown take of 6363f, and 6363g
is a breakdown take of 6363h, so 6363e likely has Grissom, and 6363g likely has
McPhail.» (July 6, 2023).