Cleaned up comments on 56-07-07 slightly
From EC:

also the Co C2K-64932 should have a * on the July9 I Got It Bad.

However, the NDESOR page doesn't list on 9-7-1956 the selection I’ve got it bad DE5614e.
The only related entry t in the Correction Sheets is (NDCS 5001):

"Session 5613. Add in NOTE: I Got It Bad was edited in the studio session 5614
to replace some Hodges' bad notes."

I would recommend you to replace the current note "Note: I Got It Bad was edited
in studio on 09 Jul 56." with the text on bold above, that is more explicit and
clear.

In conclusion, NDESOR never mentions DE5614e. And the note above doesn't refer
to a whole take of I Got It Bad on July 9. Where did Remco take the info of a
DE5614e take? Honestly, I would stay aligned yo NDESOR, removing the line
56-07-09-005 at all.