Corrected version of Hot And Bothered in DE In Order Vol 1
From EC:

according to Chuck Granata, producer of the Ellington in Order series, the
version  of "Hot And Bothered" included on Volume 1 is incorrect, as reported in
ellingtonlive facebook page on August 16.  in Ellington In Order vol.1 was
included Hot And Bothered DE3009c (12 June 1930), the same that is properly
included on vol.3, instead of the desired DE2808e (1 October 1928).  Mr Granata
promised to correct his mistake; meanwhile, I suggest to correct ellingtonia
(and ellingtonlive blog too).