Numerous corrections for 1942
From EC

1942
_______________________________________
Assumption: The RCA recordings have been updated with the 24-CD boxset Centennial edition RCA 09026-63386-2 and th3 24 LPs "Works of Duke" as per previous mails of mine.
_______________________________________
21 January

42-01-21-003	BS070683-1	C-Jam Blues DE4201c
Typo: Vi 27756 has to corrected as Vi 27856

The Moon Mist takes are currently swapped; the correct relationships takes/releases are (skip this, if you have already processed my notes about the RCA LP series "The Works of Duke"):
42-01-21-004 BS070684-1 Moon Mist (F)RCA FXM-1-7301, BL SPV-305822, Cl(F) 867 DE4201d
42-01-21-005 BS070684-2 Moon Mist Bb 5659-2-RB, Nw RP2056, Vi 27856 		 DE4201e
_______________________________________
26 February

For the following three titles
42-02-26-001	BS071890-1	What Am I Here For?	DE4202a
42-02-26-003	BS071891-2	I Don't Mind - vIA	DE4202c
42-02-26-004	BS071892-1	Someone			DE4202d
correct "Bb 5659-2" with "Bb 5659-2-RB(D)"

in the following title
42-02-26-002	BS071891-1	I Don't Mind - vIA	DE4202b
correct "Vi 09026-63386-2" with "RCA 09026-63386-2" (the RCA Centennial Edition)
_______________________________________
April / May 1942.

42-04-20-002		Barzalli Lou		DE4203b
The correct spelling of the title is Barzillai Lew.
_______________________________________
26 June

42-06-26-002	BS072438-1	Main Stem	DE4208b
also on RCA 07863-66790-2
_______________________________________
15 July
42-07-15-001	thru 42-07-15-004. These titles are issued on DETS 903-9022. Replace 903-9021 with 903-9022.
_______________________________________
28 July

** CD: Not done **
According to the RCA Centennial Edition booklet,
42-07-28-002	BS074781-2	Hayfoot, Strawfoot -vIA		DE4217b
has been issued on the 1st edition only of the 3-CD boxset Bb 5659-2-RB. The subsequent editions included 42-07-28-001, which was also issued on the European version of the boxset Bb 74321-13181-2.

The take in Bb 5659-2-RB of I Didn't Know About You (Sentimental Lady) is DE4217c
Therefore the release Bb 5659-2-RB has to be deleted from
42-07-28-003	BS074782-1 I Didn't Know About You (Sentimental Lady)	DE4217c
and added to
42-07-28-004	BS074782-2 I Didn't Know About You (Sentimental Lady)	DE4217d

42-07-28-011	Barzallai Lou
The correct spelling of this title is Barzillai Lew.
_______________________________________
13 August
Add Ivie Anderson(v) to the orchestra components. She is the singer in Massachusetts, not Joya Sherrill. The line
42-08-13-007		Massachusetts - vJS		DE4225g
has to be corrected with Massachusetts -vIA.
_______________________________________
29 September
The Duke Ellington Orchestra is also joined by a studio string orchestra.
Correct the line
42-09-29-001 Shine - vJsu, Francis Williamss, HJC
it should be
42-09-29-001 Shine - vJSu,FW,HJC
_______________________________________
October 1942 (it might be 6, 7 or 9 October)
According to DEMS Bulletin 2000/1 p.13, Duke only played three selections. The second Jubilee Theme was taken from the same recording as the first one. Therefore you can delete the line:
42-09-05-004 Jubilee Theme

In this session, the name of C. Haughton is spelled Chauncey, while in the rest of ellingtonia it is used Chauncy. Both ways are valid, but we should use only one.

_______________________________________
End of 1942____________________________