Clarified different takes of "Hayfoot, Strawfoot" on different releases of Bb 5659-2-RB
From EC:

28 July

According to the RCA Centennial Edition booklet,

42-07-28-002 BS074781-2 Hayfoot, Strawfoot -vIA DE4217b

has been issued on the 1st edition only of the 3-CD boxset Bb 5659-2-RB. The
subsequent editions included 42-07-28-001, which was also issued on the European
version of the boxset Bb 74321-13181-2.

----

Regarding the 28 July session, you are right, I didn't explain what I was asking
you to do. And now, I don't remember what I meant exactly.

The important fact is that the USA and the European version of the boxset
contained two different takes of Hayfoot, Strayfoot. But after the 2nd edition,
the USA release included the same take as the European one.

You may choose to do nothing, and leave ellingtonia as it is, or you may add a
note, like this (you probably would improve the wording):

Hayfoot, Strayfoot take 2 was released on Bb 5659-2-RB first edition only. Take
2 was included on the European release Bb 74321-13181-2 and in the subsequent
editions of Bb 5659-2-RB.