More 1943
From EC

_____________________________________
8 November (DE4359) World Transcriptions

Replace Wo 6049/6058 with 6019/6028 in the line:
43-11-08-011	Boy Meets Horn	Ci CCD-101(D),Ci CLP-101,Cl(F) 867, Wo 6049/6058 DE4359k
_____________________________________
9 November (DE4360)World Transcriptions
From now on, add violin to Ray Nance in the personnel description of all sessions till 1943/1944 (DE9071).

DeleteCi CCD-101--Ci-CLP-103 from the following line:
43-11-09-016	Things Ain't What They Used To Be	Ci CCD-101--Ci-CLP-103, Ci CCD-102(D), Ci CLP-102, Cl(F) 867, Wo 6499/6508	DE4360p

_____________________________________
1 December (DE4362) World Transcriptions

Replace TJ to Taft Jordan in the following lines
43-12-01-001	N1055-1	It Don't Mean A Thing - vRN, Taft Jordan  	DE4362a
43-12-01-002	N1055-2	It Don't Mean A Thing - vRN, Taft Jordan	DE4362b
43-12-01-003	N1055-3	It Don't Mean A Thing - vRN, Taft Jordan	DE4362c

Replace Wo 6679/6688 with Wo 6689/6698 on the line:
43-12-01-010	N1059-1	Jack The Bear	Ci CCD-103(D), Ci CLP-104, Wo 6679/6688	DE4362j

Add -vRN to
43-12-01-014	N1061-1	Ring Dem Bells	Ci CCD-103(D), Ci CLP-104, Wo 6909/6918	DE4362n

Correct Ci CLP-104 with Ci CLP-105 in the last three lines;
43-12-01-016 N1062-2 Rose Room	Ci CCD-103(D), Ci CLP-104, Wo 6689/6698		DE4362p
43-12-01-017 N1063-1 Honeysuckle Rose Ci CCD-103(D), Ci CLP-104, Wo 6679/6688 	DE4362q
43-12-01-018 N1064-1 Java Jive	Ci CCD-103(D), Ci CLP-104, Wo 6679/6688		DE4362r

^^^^^ DONE ^^^^^
_____________________________________
8 December 1943 (DE4363)
Correct Betty Rochee with Roche in the orchestra personnel.

Tracks on Sv STCD-8310 are also on the Duke Box Sv 108-8600.
_____________________________________
11 December 1943 (DE4364)

All tracks on Sv 103-8341 are also on the Duke Box Sv 108-8600.
_____________________________________
1943/1944 Unknown Location, USA
There are two entries that seem to be duplicate of the same event (indeed only one has a NDesor Id).

43-12-31-001	Do Nothin' Till You Hear From Me - vAH

44-00-00-001	Do Nothin' Till You Hear From Me - vAH		DE9071a

There are slight differences in the personnel (Harold Baker).
I believe  the first, without NDESOR ID, could be deleted.
Session DE9071 is listed in NDCS 1089, according to which personnel should probably be same as session DE4360, with Harold Baker then.

_____________________________________
End of 1943__________________________