Corrected date of DE6667 to 18 August 1966
From SE:

This time I have a question about an August 1966 session.  The SAJA
release of the "private collection, volume 10" (91234-2) has a take of
Harlem that it lists as recorded on August 18th.  You have that session
listed as August 15th, which I would normally regard as more
definitive.  However, you also show a concert on August 15th (which it
appears was a Monday) so I wonder if there is a possible typo.

(Worth noting that this same SAJA release lists The Blues as from May 6
1971 rather than May 5, but I see no reason for skepticism there.)

From EC:

You are right; August 15th is a typo. The correct date of the session is August 18th, 1966.

Regarding the session of May 1971, we find both dates, 5th and 6th, in various
sources: probably the session started on the 5th and went on till after
midnight. (I have not yet reviewed the '70s in ellingtonia.com).