Corrections for 1944
From EC:

1944
_______________________________________
Assumption: The RCA recordings have been updated with the 24 LPs "Works of Duke" as per previous mails of mine.
_______________________________________
NOTE.
At the very end of session 11 December 1943 there is a spurious "43121132" that should be deleted.
_______________________________________
1 April 1944 (DE4401)

All the selections 44-04-01-001 thru 0006 have been issued on DETS 903-9024.
_______________________________________
2 April 1944 (DE4402)
Add Al Hibbler(v) to the session personnel.
Add vAH to the line
44-04-02-010	No Love, No Nothing	DE4403h
9 April 1944 (DE4406)

Add Al Hibbler(v) to the session personnel.

Add DETS 903-9017 to all 4 selections 44-04-09-001,2,3,4

Add -vAH to the line
44-04-09-002	Do Nothin’ Till You Hear From Me	DE4406b

Replace Taft Jordan with TJ in the line
44-04-09-003	It Don't Mean A Thing - vRN, Taft Jordan DE4406c

(These two selections might be in DESUK CD 1. Pending answer from DESUK)
_______________________________________
20 April 1944 (DE4409)
21 April 1944 (DE4410)
All selections in DETS 903-9017.
_______________________________________
April 1944 (DE9089)
These selections are in DESUK CD 1
44-04-24-001	Chloe			DE9089a
44-04-24-002	Johnny Come Lately	DE9089b

_______________________________________
April 1944 (DE9090)
These selections are in DESUK CD 1
44-04-25-001	Rockabye River (Hop, Skip, Jump)		DE9090a
44-04-25-003	I Didn't Know About You (Sentimental Lady)	DE9090c
44-04-25-004	On The Alamo					DE9090d
_______________________________________
27 April 1944 (DE4412)
The three selections are in DETS 903-9018
44-04-27-001	Solid Old Man		DE4412a
44-04-27-002	How Blue The Night	DE4412b
44-04-27-003	Ring Dem Bells		DE4412c
_______________________________________
28 April 1944 (DE4413)
All selections in DETS 903-9018
44-04-28-001	Rockabye River					DE4413a
44-04-28-002	Jumping Frog Jump				DE4413b
44-04-28-003	Do Nothin' Till You Hear From Me - vAH		DE4413c
44-04-28-004	Johnny Come Lately				DE4413d
44-04-28-005	Poinciana					DE4413e
44-04-28-006	On The Alamo					DE4413f
44-04-28-007	Three Cent Stomp				DE4413g
44-04-28-008	I Didn't Know About You (Sentimental Lady)	DE4413h
_______________________________________
May 1944 (DE9094)
The selections
44-05-00-001	Rockin' In Rhythm		 DE9094a
44-05-00-002	Mood To Be Wooed		 DE9094b
are issued on DESUK CD 1

_______________________________________
May 1944 (DE9095)
All the selections
44-05-01-001	On The Alamo			 DE9095a
44-05-01-002	Main Stem			 DE9095b
44-05-01-003	My Little Brown Book - vAH	 DE9095c
44-05-01-004	Rockabye River (Hop, Skip, Jump) DE9095d
are issued on DESUK CD 1
_______________________________________
7 May 1944 (DE4416)
Replace  Taft Jordan with TJ in the line:
44-05-07-003	It Don't Mean A Thing - vRN, Taft Jordan	MJ MJCD-1124	DE4416c

_______________________________________
12 May 1944 (DE4417)
All lines in DETS 903-9018
44-05-12-001	Time Alone Will Tell		  DE4417a
44-05-12-002	San Fernando Valley		  DE4417b
44-05-12-003	Stomp, Look And Listen		  DE4417c
44-05-12-004	Concerto For Cootie		  DE4417d
_______________________________________
19 May 1944 (DE4418)
44-05-19-001	Long Ago And Far Away - vAH	  DE4418a
44-05-19-002	Someone				  DE4418b
44-05-19-003	Suddenly It Jumped		  DE4418c
44-05-19-004	Things Ain't What They Used to Be DE4418d
*** NOT YET DONE **
_______________________________________
21 May 1944 (DE4420)
Remove MJ MJCD-1140 from the line
44-05-21-009	Blue Skies	MJ MJCD-1140, MJ MJCD-1141	DE4420i
The selection is only on MJCD-1141.
_______________________________________
27 May 1944 (DE4424)
The selection
44-05-27-007	I Didn't Know About You (Sentimental Lady)	DE4424g
is also in MJ MJCD-1124
_______________________________________
4 June 1944 DE4430)
Replace Taft Jordan with TJ in the line
44-06-04-002	It Don't Mean A Thing - vRN, Taft Jordan	DE4430b
_____________________________________
8 June (DE4432)
The date is wrong, the correct date is 8 July 1944. Correct indexes 44-06-08- into 44-07-08-
The selections on Sv STCD-8310 have been issued on the Duke Box Sv 108-8600 too.

Replace  Taft Jordan with TJ in the line:
44-06-08-006	It Don't Mean A Thing - vRN, Taft Jordan	DE4432f
_____________________________________
1 December (DE4435)

Replace Kay Davis with KD in the line.
44-12-01-001	I Ain't Got Nothin' But The Blues - vAH, Kay Davis 	DE4435a
_____________________________________
17 December 1944 (DE4437)

In this session, there is possibly another title:
Do Nothin' Till You Hear From Me -vGN   AFRS YSS-86, AFRS YSS-110	DE4437c

but, as David Palmquist writes in his website, "Gertrude Niesen may not have been present, since MALB records only show her in the series on 1944-09-10 and 1945-05-27." DEMS seems to have concluded that the title was edited into the AFRS transcription disc later.
If G. Niesen was not in the same show when Ellington was guest, then her performance was not involving the Duke, and that is the probable reason why Marcus decided to ignore it (but then Gertrude Niesen should be deleted from the session personnel).
However, Do Nothin'... has a DESOR Id, DE4437c.
_____________________________________
19 December 1944 (DE4438) Carnegie Hall Concert

At the bottom of the session, replace "HBetty Roche(d)" with "Hillard Brown(d)" in the sentence:
Duke Ellington(p); Junior Raglin(sb); HBetty Roche(d) only on 44121715.
However this sentence is not fully true and could be deleted. At the very end of "Dancers In Love" there is a short fanfare with the full band.
_____________________________________
End of 1944__________________________