Added a dating note to DE3312 and corrected the date for DE3313.
From EC:
Investigating on the Ellington's 1933 sessions, I have seen that the last two
recording dates in that year were for RCA Victor, on September 26 and on
December 4.

In the booklet of the RCA 24-CD Centennial Edition, page 39, the dates reported are:

1) Midnight to 5:00 a.m. Tuesday, September 26, 1933

2) Midnight to 4:00 a.m. Monday, December 4, 1933,
with the note: the files indicate that for book-keeping purposes this session
was considered as having taken place on 12/4/33, but actually was held in the
early morning of Tuesday, 12/5/33.

There isn't such a note for the first date; therefore I understand that the
first session occurred effectively on Sept 26 and the second occurred on
December 5.  If I am not mistaken, in order to be consistent, both ellingtonia
and TDWAW should correct the date of the second session from December 4 to 5.

From SL:
Re. the December session, the files are perfectly clear that this was held in
the early morning hours of the fifth, and should therefore be shown in TDWAW
between the Regal Theatre appearances of the fourth and fifth.

The September session can be interpreted either way. The dates/times of the
adjacent matrices might clarify, but they aren't found either Brian Rust's
Victor Master Book Vol. 2, or in the DAHR (Discography of Historical Recordings,
a website).

RCA's session sheet for the first session shows the date September 26. 1933 and
"Time 12:00 to 5:00." Below and above 12:00 appears "A.M."; "A.M" also appears
below 5:00.

The 1933 09 26 sheet also notes "Mr. Young -- present." This was someone other
than Victor Young, who worked in NYC for ARC/Brunswick throughout 1933.