Correct vocalist on DE4436c to Joya Sherrill (was Betty Roche). With thanks to Stephen Macatee.
From SM:

The 11, 12 December 1944 Discography page lists “Betty Roche” (-vBR) as the
vocalist on the song The Blues. But the physical 78 RPM disc lists the singer as
Joya Sherrill.

From EC:

here Stephen has caught a mistake that I had not seen.
Indeed Joya Sherrill is the singer on The Blues on 11-12 December 1944.
Therefore on the DE4436 session,  Charlie should replace, please, Betty Roche(v)
with Joya Sherrill(v) in the band personnel, and - vBR with - JS in the line:
44-12-11-003.    D4VC-562-3      06. The Blues - vBR