Various 1973 fixes
From EC:

10 January 1973 (DE7306)
Location: Schubert Theatre.

Replace the description of the orchestra, as per NDESOR:

7 trumpets, 5 trombones, 3 horns, 8 reeds, 12 violins, 4 violas, 4 cellos, 1 harp, 9 rhythms, cond. by Quincy Jones. Among them....
In the list of names, replace Booty Wood(tb) with Britt Woodman in the orchestra.
Do not repeat Quincy Jones at the end.

After DE7306c Duke Ellington(tk) makes the Finger Snapping Routine, accompanied by the orchestra that plays "Love You Madly" theme in background. It is not mentioned in ellingtonia.
_______________________________________
31 January 1973 (DE7310) II set

Please insert the following line:

73-01-31-028	Don't You Know I Care 		DE7310i

after the line:
73-01-31-027	Take The "A" Train		DE7310h

As a consequence, indexes and NDesor IDs have to be increased, starting from
73-01-31-029	Fife				DE7310j

The list line will be then:
73-01-31-040	Things Ain't What They Used To Be	DE7310u

_______________________________________
01 February 1973 (DE7311) I set

Please insert the following line:
73-02-01-014	Sophisticated Lady		DE7311n

after the line:
73-02-01-013	Hello, Dolly! - vMJ		DE7311m

As a consequence, indexes and NDesor IDs have to be increased.