Added missing track in 73-04-12 concert; minor corrections
From EC:

12 April 1973 (DE7322)
Add "Concert", next to Queen Elizabeth Theatre.

Delete the  NDESOR ID from the line
73-04-12-023	4. Sophisticated Lady			DE7322t

So that in the subsequent lines, the NDESOR IDs are wrong; they should be:
73-04-12-024	Making That Scene - vTW			DE7322t
73-04-12-025	Do Nothin' Till You Hear From Me - vTW	DE7322u
73-04-12-026	Solitude - vTW				DE7322v
73-04-12-027	Mood Indigo - vTW			DE7322w
73-04-12-028	One More Time For The People - vTW	DE7322x

One title is missing at this point. Please add the line:
73-04-12-029	Things Ain't What They Used To Be	DE7322y

After that, NDESOR IDs are correct (line indexes need to be updated).