Many 1973 fixes
From EC:

11 May 1973 (DE7327)
The city is Kenosha, not Kenosh.

The title Woods should be replaced with Soso:
73-05-11-008	Woods				DE7327h

The title in the following line is wrong:
73-05-11-016	Piano Interlude			DE7327p
it is
73-05-11-016	I Let A Song Go Out Of My Heart DE7327p

_______________________________________
25 May 1973 (DE7328)

Add clarinet to Russell Procope in the line preceding "Tiger Rag"; it should be:

Harold "Money" Johnson(t); Vincente Prudente(tb); Russell Procope(as,cl); Duke Ellington(p); Joe Benjamin(sb); Rufus Jones(d)
73-05-25-023		Tiger Rag		DE7328xm

_______________________________________
03 July 1973 (DE7329)

Emmanuel Abdul-Rahim(cg) instead of Aziz Lateef(cg) in the band.
_______________________________________
16 July 1973 (DE7331)
City and location are wrong (copied from previous date).

City: Skokie, Illinois
Location: Concert. Old Orchard Shopping Center

Add (theme) to the initial Take The A Train.
_______________________________________
21 July 1973 (DE7332)

The date seems to be incorrect: a ticket, posted on tdwaw. shows the date 20 July 1973.

As usual, Woods should be replaced with Soso in the line:
73-07-21-012		Woods				DE7332l

Replace Piano Interlude with Ring Dem Bells in the line:
73-07-21-019		Piano Interlude			DE7332s

Add clarinet to Russell Procope in the group performing Tiger Rag and St. Louis Blues (lines 020, 021)

Replace -vTW with -vMJ in the line:
73-07-21-028		Basin Street Blues - vTW	DE7332ab

Add Q to the unidentified below:
73-07-21-037		9. Unidentified

_______________________________________
31 July 1973 (DE7333)

Emmanuel Abdul-Rahim(cg) instead of Aziz Lateef(cg) in the band.

As usual, Woods should be replaced with Soso in the line:
73-07-31-006		Woods		DE7333f

_______________________________________
03 August 1973 (DE7334)

Emmanuel Abdul-Rahim(cg) instead of Aziz Lateef(cg) in the band.

Woods should be replaced with Soso in the line:
73-08-03-023		Woods		DE7334w

_______________________________________
06 August 1973 (DE7335)

The location is wrong, copied from a previous date, This is the correct description:
NBC Telecast, titled "Today Show". NBC Studios.

Emmanuel Abdul-Rahim(cg) instead of Aziz Lateef(cg) in the band.  (Check on NDESOR if he was really present).

Woods should be replaced with Soso in the line:
73-08-06-002		Woods		DE7335b

_______________________________________
August 1973 (DE7336)

Swedish TV broadcast.

In the band, replace Mercer Ellington with Harold "Money" Johnson.

Emmanuel Abdul-Rahim(cg) instead of Aziz Lateef(cg) in the band.

_______________________________________
25 August 1973 (DE7337)
Add Toney Watkins to the band.