Re-dated and moved DE7341 with minor tweaks
From EC
_______________________________________
25 August 1973 (DE7341)
This session is in the wrong position, due to a wrong dating.

The right date is 20 September 1973, and the right position is after session DE7342 (6 September). Of course, all indexes must be changed into 73-09-20-001 and so on.

Add Toney Watkins(v) to the band.

In the following lines, add -vRN and remove (theme):
73-08-25-055		Take The "A" Train (theme)	DE7341n
73-08-25-056		Just A-Sittin' And A-Rockin'	DE7341o

Change "Over the Waves" into "Anniversary Song" (DEMS 04/1-32, DESOR small corrections 5007):
73-08-25-064		Over The Waves			DE7341w