More 1973
From EC:
_______________________________________
14 November 1973 (DE7354)

Add Sq SR-0197 to
73-11-14-014		Hello, Dolly! - vMJ		DE7354n

Woods should be replaced with Soso in the line:
73-11-14-023		Woods	Sq SR-0197		DE7354q

In NDESOR there is this note:
NOTE - Hallelujah is played by Duke Ellington, Joe Benjamin and Quentin White only.
_______________________________________
16 November 1973 (DE7355)

All selections have been issued on the DVD "Duke Ellington and His Orchestra Live", EuroArts 2064938.

_______________________________________
18 November 1973 (DE9005)

The following line is missing,
73-11-18-005		Spacemen			DE9005e

between
73-11-18-004		Creole Love Call		DE9005d
and
73-11-18-005		In A Sentimental Mood		DE9005f
(index has to be increased tp 73-11-18-006, and so on for the following lines.)

Remove Woods and leave only Soso in the line:
73-11-18-025		Woods (Soso)		DE9005w

______________________________________
26 November 1973 (DE7356)

The concert was a charity show, its title is "Royal Command Variety Performance".

_____________________________________
30 November 1973 (DE9001)

Woods should be replaced with Soso in the line:
73-11-30-025		Woods		DE9001u

______________________________________
01 December 1973 (DE7357)

Woods should be replaced with Soso in the line:
73-12-01-026		Woods			DE7357w