Fixed description for DE5836
From DP:

How about “Filming session for future telecast (aired 1959-01-10)”

i only suggest the above wording because the shoot wasn’t finished until
Mitchell went home to practice fingering in synch with the soundtrack so he
could be filmed “playing” in the episode.

If I understand it correctly, unlike video recording, motion picture camera
technology required the musicians to fake it on camera, listening to playback of
the music they’d prerecorded.  This was because the camera angles would change,
or more than one camera would be used, but unlike video, the film could not be
mixed.

I am pretty sure this was filmed, not taped, because one of the newspapers
printed a picture of Formby being filmed on the ship. The camera was on a tripod
about 5 or 6 feet from his face, and I recall it had typlcal film canisters on
top.  Of course I cannot find that picture again, and my memory may not be
great!