1960 corrections
Rreviewing your corrections to 1960.
As per 1972, the location should not have been added to the city, in the session header, but in the first line, next to the event description.
It means a significant amount of rework.

This is the case of
1) 21 February 1960 (DE6002). Location: William and Mary College.
It should have been added after WCWN broadcast: WCWN broadcast William and Mary College.
Remove Willian and Mary College next to the city: Willian and Mary College, Williamsburg, VA.

2) Correction omitted.
26 May 1960 (DE6006). Location: Radio Recorders.
Add Radio Recorders after Columbia Recording Session.

Exactly the same correction on
31 May 1960 (DE6008)
01 June 1960 (DE6009)
02 June 1960 (DE6010)
03 June 1960 (DE6011)

3) Move Radio Recorders from the header "Radio Recorders, Los Angeles, CA" to the description "Columbia Recording Session. Radio Recorders."
for the dates: 20, 21, 22, 26, 28, 29, 30 June; 1, 14 July; 10 October.

4) Correction omitted.
04 October 1960 (DE6031). Correct the description: KHIP broadcast. Club Névé.

5) 14 December. Location Barclay Studio, to be moved from the header (Barclay Studio, Paris, France) to the description, indeed missing.
Therefore add the line:
Recording session. Barclay Studio.
before the line Duke Ellington(p).
Remove the whole Barclay Studio, Paris, France from the header of the second part of the session (after the first 4 lines).

6) 17 December. MY MISTAKE, the show is "Jean Sablon Show", not Juan. Sorry.